+++ src/config/DeviceConfig.cpp
@@ -1,6 +1,6 @@
 #include "config/DeviceConfig.hpp"
 
-DeviceConfig::DeviceConfig(uint16_t _id, const char* _name, DeviceType _type, uint8_t _portNumber, WireColor _wire, uint16_t _debounceMs):
+DeviceConfig::DeviceConfig(uint8_t _id, const char* _name, DeviceType _type, uint8_t _portNumber, WireColor _wire, uint16_t _debounceMs):
     id(_id), 
     name(_name),
     type(_type),
